+++ rpc/src/endpoint/genesis.rs
@@ -16,6 +16,8 @@ impl crate::Request for Request {
     }
 }
 
+impl crate::SimpleRequest for Request {}
+
 /// Block responses
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Response {