+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ### IMPROVEMENTS:
 
 - `[light-client]` Only require Tokio when `rpc-client` feature is enabled ([#425])
+- `[rpc]` The `WebSocketClient` now adds support for all remaining RPC requests
+  by way of implementing the `Client` trait ([#646])
 - `[rpc]` Support for the `tx_search` RPC endpoint has been added ([#701])
 
 [#425]: https://github.com/informalsystems/tendermint-rs/issues/425
+[#646]: https://github.com/informalsystems/tendermint-rs/pull/646
 [#701]: https://github.com/informalsystems/tendermint-rs/pull/701
 
 ## v0.17.0-rc3